refactor(studio): migrate schedules router onto registry (phase 2 area 3)
Move all /schedules/* handlers from routers/schedules.py into
services/schedules.py decorated with @studio_route. Request/response
models (CreateScheduleRequest, UpdateScheduleRequest) are co-located in
the service file. Route handlers are renamed with _route suffix to avoid
collision with service functions; name= overrides preserve original
FastAPI route names.

The get_schedule_run route uses tags=["schedules", "schedule-runs"] to
reproduce the original two-tag assignment from APIRouter(tags=["schedules"])
plus @router.get(..., tags=["schedule-runs"]).

Oracle: routes.json identical, openapi.json identical (79 routes).
