+++ lionagi/studio/registry.py
@@ -60,6 +60,7 @@ _STUDIO_ROUTE_MODULES: tuple[str, ...] = (
     "lionagi.studio.services.engine_defs",
     "lionagi.studio.services.sessions",
     "lionagi.studio.services.admin",
+    "lionagi.studio.services.schedules",
 )
